Organize all the sub-modes of the DesignLTN page
