+++ apps/ltn/src/pages/design_ltn/shortcuts.rs
