+++ apps/ltn/src/render/mod.rs
@@ -6,7 +6,8 @@ use std::collections::HashSet;
 use geom::Distance;
 use map_model::osm::RoadRank;
 use map_model::{AmenityType, Map, RoadID};
-use widgetry::{Color, Drawable, EventCtx, GeomBatch, RewriteColor};
+use widgetry::mapspace::DrawCustomUnzoomedShapes;
+use widgetry::{Color, Drawable, EventCtx, GeomBatch, GfxCtx, RewriteColor};
 
 use crate::App;
 
@@ -94,4 +95,30 @@ pub fn render_bus_routes(ctx: &EventCtx, map: &Map) -> Drawable {
         }
     }
     ctx.upload(batch)
+}
+
+/// Depending on the canvas zoom level, draws one of 2 things.
+// TODO Rethink filter styles and do something better than this.
+pub struct Toggle3Zoomed {
+    draw_zoomed: Drawable,
+    unzoomed: DrawCustomUnzoomedShapes,
+}
+
+impl Toggle3Zoomed {
+    pub fn new(draw_zoomed: Drawable, unzoomed: DrawCustomUnzoomedShapes) -> Self {
+        Self {
+            draw_zoomed,
+            unzoomed,
+        }
+    }
+
+    pub fn empty(ctx: &EventCtx) -> Self {
+        Self::new(Drawable::empty(ctx), DrawCustomUnzoomedShapes::empty())
+    }
+
+    pub fn draw(&self, g: &mut GfxCtx) {
+        if !self.unzoomed.maybe_draw(g) {
+            self.draw_zoomed.draw(g);
+        }
+    }
 }
